requirements.txt
streamlit==1.35.0
pandas==2.2.2
mysql-connector-python==8.3.0
scikit-learn==1.4.2
numpy==1.26.4
